Converted to a module (#1)
