+++ replit.nix
@@ -1,6 +1,7 @@
 { pkgs }: {
 	deps = [
 		pkgs.python38Full
+		pkgs.neovim
 	];
   env = {
     LD_LIBRARY_PATH = pkgs.lib.makeLibraryPath ([
